requirements.txt
Django==3.0.6
djangorestframework==3.11.0
djoser==2.0.3
